[tf.data] move map, mad_and_batch serialization tests to kernel_tests
